requirements.txt
appnope==0.1.0
argh==0.26.2
astroid==2.0.4
atomicwrites==1.1.5
attrs==18.1.0
backcall==0.1.0
certifi==2018.8.13
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
decorator==4.3.0
Django==2.1
django-filter==2.0.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
docopt==0.6.2
idna==2.7
ipython==6.5.0
ipython-genutils==0.2.0
isort==4.3.4
itypes==1.1.0
jedi==0.12.1
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
openapi-codec==1.3.2
parso==0.3.1
pathtools==0.1.2
pexpect==4.6.0
pickleshare==0.7.4
pluggy==0.7.1
prompt-toolkit==1.0.15
psycopg2-binary==2.7.5
ptyprocess==0.6.0
py==1.5.4
Pygments==2.2.0
pylint==2.1.1
pytest==3.7.1
pytest-django==3.3.3
pytest-watch==4.2.0
pytz==2018.5
PyYAML==3.13
requests==2.19.1
simplegeneric==0.8.1
simplejson==3.16.0
six==1.11.0
traitlets==4.3.2
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.23
watchdog==0.8.3
wcwidth==0.1.7
wrapt==1.10.11
